Move CKZG property tests to :ethereum:spec
