+++ infrastructure/kzg/build.gradle
@@ -6,11 +6,4 @@ dependencies {
   implementation 'org.apache.tuweni:tuweni-ssz'
   implementation 'tech.pegasys:jc-kzg-4844'
   implementation 'commons-io:commons-io'
-
-  testImplementation project(':ethereum:spec')
-  testImplementation testFixtures(project(':ethereum:spec'))
-
-  testFixturesImplementation 'net.jqwik:jqwik'
-  testFixturesImplementation 'org.apache.tuweni:tuweni-bytes'
-  testFixturesImplementation testFixtures(project(':ethereum:spec'))
 }
